% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AH_AOC.R
\name{AH_AOC}
\alias{AH_AOC}
\title{Calculate Area Outside the Curve}
\usage{
AH_AOC(dat, input, output, method, print_style = "ascii", CI = FALSE,
  nboots)
}
\arguments{
\item{dat}{A Dataset}

\item{input}{Independent variable}

\item{output}{Dependent variable}

\item{method}{What boundary method should be used? One of: SFA, QR, Polynomial, Kernel.}

\item{print_style}{"ascii" or "r"?}

\item{CI}{Should confidence intervals be returned?}

\item{nboots}{Number of bootstraps to generate the confidence intervals}
}
\description{
\code{AH_AOC} calculates the Area Outside of the Curve for asymmetric boundaries. It is implemented for a variety of techniques and can return bootstrapped confidence intervals.
}
\examples{
DATA <- read.csv("dat.csv")
AH_AOC(DATA, input = DATA$x, output = DATA$y, method = "QR",
   print_style = "ascii", CI = TRUE, nboots = 5000)
}

