% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_boundary.R
\name{estimate_boundary}
\alias{estimate_boundary}
\title{Estimate Boundary of Asymmetric Hypothesis}
\usage{
estimate_boundary(dat, input, output, sufficient = FALSE, method,
  AOC = FALSE)
}
\arguments{
\item{dat}{A Dataset}

\item{input}{Independent variable}

\item{output}{Dependent variable}

\item{sufficient}{Is the data sufficient? Default is FALSE.}

\item{method}{What boundary technique? One of: SFA, QR, Polynomial, Kernel.}

\item{AOC}{Should AOC be returned? Default is FALSE.}
}
\description{
This function uses a variety of techniques to estimate data boundaries for asymmetric hypotheses.
}
\examples{
DATA <- read.csv("dat.csv")
estimate_boundary(dat = DATA, input = DATA$x, output = DATA$y, sufficient = FALSE, merthod = "QR", AOC = FALSE)
}

