% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{AH_AOC}
\alias{AH_AOC}
\title{Calculate Area Outside the Curve}
\usage{
AH_AOC(dat, input, output, method, print_style = "ascii", CI = FALSE,
  nboots)
}
\arguments{
\item{dat}{A Dataset}

\item{input}{Independent variable}

\item{output}{Dependent variable}

\item{method}{What boundary method should be used?}

\item{print_style}{"ascii" or "r"?}

\item{CI}{Should confidence intervals be returned?}

\item{nboots}{Number of bootstraps to generate the confidence intervals}
}
\description{
Calculate Area Outside the Curve
}

